Move UserFeed components to their own directory
